Power counter experiments
This explores an idea that we discussed earlier on IRC, to change
the displayed power value after taking into account the production
icon that is under the mouse pointer.

It didn't turn out as good as I had hoped. I tried a few different
things but none was great.

There are two slightly different demos in this PR, try them both
by changing the #define at the top of IngameSiderbarComponentsLogic.cs
to TEST1 or TEST2. Make, run, and hover over the production icons
to see them in action.